DRY in checkbox filters, consolidate their folder structure
