+++ source/Playnite/Settings/Filters/FullscreenFilterSettings.cs
