+++ source/Playnite/Settings/Filters/ThreeStateFilterEnum.cs
